made teleporter trigger modular
