+++ Assets/Scripts/PickupItem.cs
@@ -21,4 +21,8 @@ public class PickupItem : MonoBehaviour
         this.gameObject.SetActive( true );
         this.gameObject.transform.position = dropPostion;
     }
+
+    public void OnPlayerLooking()
+    {
+    }
 }
