+++ Assets/Scripts/TeleporterTrigger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ab25250fc2396ad4cb1ae9054b540845
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
